pyproject.toml
[project]
name = "AutoNews"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fake-useragent>=2.2.0",
    "openai>=1.90.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "selenium>=4.33.0",
    "selenium-stealth>=1.0.6",
    "seleniumbase>=4.39.5",
    "setuptools>=80.9.0",
    "undetected-chromedriver>=3.5.5",
    "webdriver-manager>=4.0.2",
    "yt-dlp>=2025.6.30",
]
